+++ app/server/tests/unit/controllers/projectController.test.ts
@@ -8,48 +8,27 @@ const mockProjectSamples = [
 const mockUserStore = {
     saveNewProject: jest.fn().mockImplementation(() => "test-project-id"),
     getUserProjects: jest.fn().mockImplementation(() => mockUserProjects),
-    saveProjectHash: jest.fn(),
     getProjectHash: jest.fn().mockImplementation(() => "123"),
     saveAMR: jest.fn(),
     getProjectSamples: jest.fn().mockImplementation(() => mockProjectSamples),
     getAMR: jest.fn().mockImplementation((projectId: string, sampleHash: string, fileName: string) =>
         `AMR for ${projectId}-${sampleHash}-${fileName}`)
 };
-jest.mock("../../src/db/userStore", () => ({
+jest.mock("../../../src/db/userStore", () => ({
     userStore: mockUserStoreConstructor.mockReturnValue(mockUserStore)
-}))
+}));
 
-import axios from 'axios';
-import MockAdapter from 'axios-mock-adapter';
-import {apiEndpoints} from '../../src/routes/routes';
-import versionInfo from '../../../server/resources/versionInfo.json';
-import config from '../../src/resources/config.json';
+import {mockApp, mockRedis, mockResponse} from "../utils";
+import config from "../../../src/resources/config.json";
+import projectController from "../../../src/controllers/projectController";
+import MockAdapter from "axios-mock-adapter";
+import axios from "axios";
 
-import {mockResponse} from "./utils";
+const mockAxios = new MockAdapter(axios);
 
-const mockRequest: any = { };
-
-const mockRedis = {};
-
-const mockApp = {
-    locals: {
-        redis: mockRedis
-    }
-};
-
-describe("test routes", () => {
-    const mockAxios = new MockAdapter(axios);
-    mockAxios.onGet(`${config.api_url}/version`).reply(200, versionInfo);
-
+describe("projectController", () => {
     beforeEach(() => {
         jest.clearAllMocks();
-    });
-
-    it("get version info", async () => {
-        const req = mockRequest;
-        const res = mockResponse();
-        await apiEndpoints(config).getVersionInfo(req, res);
-        expect(res.send).toHaveBeenCalledWith(versionInfo)
     });
 
     it("saves new project", async () => {
@@ -60,7 +39,7 @@ describe("test routes", () => {
             app: mockApp
         };
         const res = mockResponse();
-        await apiEndpoints(config).newProject(req, res);
+        await projectController(config).newProject(req, res);
         expect(mockUserStoreConstructor).toHaveBeenCalledTimes(1);
         expect(mockUserStoreConstructor.mock.calls[0][0]).toBe(mockRedis);
         expect(mockUserStore.saveNewProject).toHaveBeenCalledTimes(1);
@@ -74,45 +53,12 @@ describe("test routes", () => {
         });
     });
 
-    it("sets hash for project and forwards request when run poppunk", async () => {
-        const expectedPoppunkReq = {
-            projectHash: "1234",
-            names: {
-                sample1: "file1.fasta",
-                sample2: "file2.fasta"
-            },
-            sketches: {
-                sample1: {7: "abcd"},
-                sample2: {7: "efgh"}
-            }
-        };
-
-        const req = {
-            body: {
-               ...expectedPoppunkReq,
-                projectId: "test-project-id"
-            },
-            app: mockApp
-        };
-
-        await apiEndpoints(config).runPoppunk(req, {}, jest.fn());
-        expect(mockUserStoreConstructor).toHaveBeenCalledTimes(1);
-        expect(mockUserStoreConstructor.mock.calls[0][0]).toBe(mockRedis);
-        expect(mockUserStore.saveProjectHash).toHaveBeenCalledTimes(1);
-        expect(mockUserStore.saveProjectHash.mock.calls[0][0]).toBe(req);
-        expect(mockUserStore.saveProjectHash.mock.calls[0][1]).toBe("test-project-id")
-        expect(mockUserStore.saveProjectHash.mock.calls[0][2]).toBe("1234");
-
-        expect(mockAxios.history.post[0].url).toBe("http://localhost:5000/poppunk");
-        expect(JSON.parse(mockAxios.history.post[0].data)).toStrictEqual(expectedPoppunkReq);
-    });
-
     it("gets projects for user", async () => {
         const req = {
             app: mockApp
         };
         const res = mockResponse();
-        await apiEndpoints(config).getProjects(req, res, jest.fn());
+        await projectController(config).getProjects(req, res, jest.fn());
         expect(mockUserStoreConstructor).toHaveBeenCalledTimes(1);
         expect(mockUserStoreConstructor.mock.calls[0][0]).toBe(mockRedis);
         expect(mockUserStore.getUserProjects).toHaveBeenCalledTimes(1);
@@ -127,18 +73,18 @@ describe("test routes", () => {
 
     it("saved amr data", async () => {
         const req = {
-           app: mockApp,
-           body: {
-               filename: "test.fa",
-               Penicillin: 0.5
-           },
-           params: {
-               projectId: "testProjectId",
-               sampleHash: "1234"
-           }
+            app: mockApp,
+            body: {
+                filename: "test.fa",
+                Penicillin: 0.5
+            },
+            params: {
+                projectId: "testProjectId",
+                sampleHash: "1234"
+            }
         };
         const res = mockResponse();
-        await apiEndpoints(config).postAMR(req, res, jest.fn());
+        await projectController(config).postAMR(req, res, jest.fn());
         expect(mockUserStoreConstructor.mock.calls[0][0]).toBe(mockRedis);
         expect(mockUserStore.saveAMR).toHaveBeenCalledWith("testProjectId", "1234", req.body);
     });
@@ -160,7 +106,7 @@ describe("test routes", () => {
 
         mockAxios.onGet(`${config.api_url}/project/123`).reply(200, {data: projectData});
 
-        await apiEndpoints(config).getProject(req, res, jest.fn());
+        await projectController(config).getProject(req, res, jest.fn());
 
         const response = res.json.mock.calls[0][0];
         expect(response).toStrictEqual({
@@ -196,7 +142,7 @@ describe("test routes", () => {
         mockAxios.onGet(`${config.api_url}/project/123`).reply(200, {data: projectData});
         const next = jest.fn();
 
-        await apiEndpoints(config).getProject(req, res, next);
+        await projectController(config).getProject(req, res, next);
         expect(next).toHaveBeenCalledTimes(1);
         expect(next.mock.calls[0][0].message).toBe("Sample with hash 1234 was not in API response");
     });
@@ -219,7 +165,7 @@ describe("test routes", () => {
         mockAxios.onGet(`${config.api_url}/project/123`).reply(500, mockError);
 
         const next = jest.fn();
-        await apiEndpoints(config).getProject(req, res, next);
+        await projectController(config).getProject(req, res, next);
 
         expect(res.status).toHaveBeenCalledWith(500);
         expect(res.json).toHaveBeenCalledWith({
